Move HmdaFileUpload to submission package
